Last Update

+++ GeoCity/client/src/app/services/auth.service.ts
@@ -1,11 +1,15 @@
-import { Injectable } from '@angular/core';
+import { Injectable, OnDestroy } from '@angular/core';
 import { Http } from '@angular/http';
+import { Router } from '@angular/router';
 import 'rxjs/add/operator/map';
 
 @Injectable()
 export class AuthService {
 
-    constructor(public http: Http) {
+    constructor(
+        public http: Http,
+        public router: Router
+    ) {
     }
 
     login(user) {
@@ -28,21 +32,36 @@ export class AuthService {
         return false;
     }
 
-    logout(){
-        return this.http.post('http://localhost:3000/api/Users/logout?access_token=' + localStorage.getItem('token'),localStorage.getItem('token'))
-        .map(res => {
-            localStorage.clear();
-        });
-    }
-
-    resigter(user){
-        return this.http.post('http://localhost:3000/api/Clients',user)
-        .map(res => {
-            res.json();
-            console.log(res.json());
-        })
-    }
-
-
+    logout() {
+        if (localStorage.getItem('token') !== null) {
+            return this.http.post('http://localhost:3000/api/Users/logout?access_token=' + localStorage.getItem('token'), localStorage.getItem('token'))
+                .map(res => {
+                    localStorage.clear();
+                });
+        } else {
+            this.router.navigate(['/login']);
+        }
 }
 
+    resigter(user) {
+        return this.http.post('http://localhost:3000/api/Clients', user)
+            .map(res => {
+                res.json();
+                console.log(res.json());
+            })
+    }
+
+    getRole(id) {
+        return this.http.get('http://localhost:3000/api/Clients/getRole?id=' + localStorage.getItem('userID'))
+            .map(res => {
+                return res.json();
+            })
+    }
+
+    isEmployee() {
+        if (localStorage.getItem('role') === 'Employee'.toLowerCase()) {
+            return true;
+        }
+        return false;
+    }
+}
